feat: create CardIndicator component and use it in InfoBubble component
