+++ packages/ui/src/components/molecules/index.tsx
@@ -22,6 +22,7 @@ export {default as Badge} from './Badge/Badge';
 export {default as BlockInteractionMessage} from './BlockInteractionMessage/BlockInteractionMessage';
 export {default as Button} from './Button/Button';
 export {default as CardIconButton} from './CardIconButton/CardIconButton';
+export {default as CardIndicator} from './CardIconButton/CardIconButton';
 export {default as Checkbox} from './Checkbox/Checkbox';
 export {default as Chip} from './Chip/Chip';
 export {default as CircleButton} from './CircleButton/CircleButton';
@@ -38,7 +39,6 @@ export {default as HalfLabelCard} from './HalfLabelCard/HalfLabelCard';
 export {default as IconInput} from './IconInput/IconInput';
 export {default as Image} from './Image/Image';
 export {default as Increment} from './Increment/Increment';
-export {default as InfoBubble} from './InfoBubble/InfoBubble';
 export {default as Label} from './Label/Label';
 export {default as LabelText} from './LabelText/LabelText';
 export {default as LoadingIndicator} from './LoadingIndicator/LoadingIndicator';
