+++ packages/ui/src/components/organisms/GroupByScrollList/TopSeparator.tsx
@@ -19,7 +19,8 @@
 import React, {useMemo} from 'react';
 import {ActivityIndicator, Dimensions, StyleSheet, View} from 'react-native';
 import {Text} from '../../atoms';
-import {InfoBubble, NumberBubble} from '../../molecules';
+import {NumberBubble} from '../../molecules';
+import InfoBubble from '../InfoBubble/InfoBubble';
 import {Color, useThemeColor} from '../../../theme';
 
 export interface TopIndicator {
