+++ packages/ui/src/components/organisms/InfoBubble/InfoBubble.tsx
@@ -1,0 +1,79 @@
+/*
+ * Axelor Business Solutions
+ *
+ * Copyright (C) 2024 Axelor (<http://axelor.com>).
+ *
+ * This program is free software: you can redistribute it and/or  modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import React, {useMemo} from 'react';
+import {Dimensions, StyleSheet} from 'react-native';
+import {Icon} from '../../atoms';
+import {Color} from '../../../theme/themes';
+import CardIndicator from '.././../molecules/CardIndicator/CardIndicator';
+
+interface InfoBubbleProps {
+  style?: any;
+  textIndicationStyle?: any;
+  iconName: string;
+  badgeColor: Color;
+  indication: string;
+  size?: number;
+  position?: 'left' | 'right';
+  coloredBubble?: boolean;
+}
+
+const InfoBubble = ({
+  style,
+  iconName,
+  badgeColor,
+  indication,
+  textIndicationStyle,
+  size = Dimensions.get('window').width * 0.07,
+  position = 'right',
+  coloredBubble = true,
+}: InfoBubbleProps) => {
+  const styles = useMemo(() => getStyles(badgeColor, size), [badgeColor, size]);
+
+  return (
+    <CardIndicator
+      style={style}
+      indication={indication}
+      position={position}
+      textIndicationStyle={textIndicationStyle}>
+      <Icon
+        name={iconName}
+        style={coloredBubble ? styles.icon : null}
+        color={coloredBubble ? badgeColor.foreground : badgeColor.background}
+        size={coloredBubble ? size * 0.5 : size * 0.8}
+      />
+    </CardIndicator>
+  );
+};
+
+const getStyles = (badgeColor, size) =>
+  StyleSheet.create({
+    icon: {
+      alignSelf: 'center',
+      justifyContent: 'center',
+      alignItems: 'center',
+      backgroundColor: badgeColor.background_light,
+      borderWidth: 2,
+      borderColor: badgeColor.background,
+      borderRadius: size,
+      width: size,
+      height: size,
+    },
+  });
+
+export default InfoBubble;
